# Improve numerical stability of GroupNorm