--- a/aten/src/ATen/native/cpu/utils.h
+++ b/aten/src/ATen/native/cpu/utils.h
@@ -3,7 +3,10 @@
 #include <ATen/cpu/vec/vec.h>
 #include <c10/util/llvmMathExtras.h>
 
-namespace at { namespace native { namespace {
+namespace at {
+namespace native {
+
+namespace {
 
 template <typename T>
 inline T data_index_init(T offset) {
@@ -11,7 +14,7 @@ inline T data_index_init(T offset) {
 }
 
 template <typename T, typename... Args>
-inline T data_index_init(T offset, T &x, const T &X, Args &&... args) {
+inline T data_index_init(T offset, T& x, const T& X, Args&&... args) {
   offset = data_index_init(offset, std::forward<Args>(args)...);
   x = offset % X;
   return offset / X;
@@ -22,7 +25,7 @@ inline bool data_index_step() {
 }
 
 template <typename T, typename... Args>
-inline bool data_index_step(T &x, const T &X, Args &&... args) {
+inline bool data_index_step(T& x, const T& X, Args&&... args) {
   if (data_index_step(std::forward<Args>(args)...)) {
     x = ((x + 1) == X) ? 0 : (x + 1);
     return x == 0;
@@ -47,4 +50,4 @@ T CeilLog2(const T& x) {
 } // namespace utils
 
 } // namespace native
-} // namespace at// namespace at::native::<anonymous>
+} // namespace at