Summary: Improve numerical stability of GroupNorm

Test Plan: buck test mode/dev-nosan //caffe2/test:nn -- "GroupNorm"

Differential Revision: D27414438

There are two main changes in this PR:
1. Use Welford algorithm to compute main and variance.
2. Use cascade summation when computing sum over input for both mean and variance.

Some benchmark results.

input_size = [256, 1024, 28, 28], group = 8

for single thread CPU run average time change is 510.91ms -> 602.70ms.
for CUDA run average time change is 12.20ms -> 12.47ms.